Extracting model classes
