+++ libs/support/src/main/java/org/wordpress/android/support/feature/aibot/ConversationUtils.kt
@@ -3,6 +3,8 @@ package org.wordpress.android.support.feature.aibot
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalContext
 import org.wordpress.android.support.R
+import org.wordpress.android.support.feature.aibot.model.BotConversation
+import org.wordpress.android.support.feature.aibot.model.BotMessage
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
