+++ libs/support/src/main/java/org/wordpress/android/support/feature/aibot/model/BotConversation.kt
@@ -1,0 +1,11 @@
+package org.wordpress.android.support.feature.aibot.model
+
+import java.util.Date
+
+data class BotConversation(
+    val id: Long,
+    val createdAt: Date,
+    val mostRecentMessageDate: Date,
+    val lastMessage: String,
+    val messages: List<BotMessage>
+)
