+++ libs/support/src/main/java/org/wordpress/android/support/feature/aibot/model/BotMessage.kt
@@ -1,0 +1,11 @@
+package org.wordpress.android.support.feature.aibot.model
+
+import java.util.Date
+
+data class BotMessage(
+    val id: Long,
+    val text: String,
+    val date: Date,
+    val userWantsToTalkToHuman: Boolean,
+    val isWrittenByUser: Boolean
+)
